Fixed bug

+++ main.py
@@ -18,18 +18,19 @@ BotWrapper.startBot()
 while True:
 
     out = sender.check()
-
-    for mail in out:
-        out_string = "*Neuer Druckauftrag*\n\n"
-        for key in mail:
-            out_string += "_"+key+"_"
-            out_string += "\n   "
-            out_string += mail[key]
-            out_string += "\n"
-        BotWrapper.sendMessage(config.chatID, out_string)
-        if "file" in mail:
-            BotWrapper.botBackend.dispatcher.bot.send_document(
-                chat_id=config.chatID, document=open(mail['file'], 'rb'))
-            os.remove(mail['file'])
+    if out is not None:
+        
+        for mail in out:
+            out_string = "*Neuer Druckauftrag*\n\n"
+            for key in mail:
+                out_string += "_"+key+"_"
+                out_string += "\n   "
+                out_string += mail[key]
+                out_string += "\n"
+            BotWrapper.sendMessage(config.chatID, out_string)
+            if "file" in mail:
+                BotWrapper.botBackend.dispatcher.bot.send_document(
+                    chat_id=config.chatID, document=open(mail['file'], 'rb'))
+                os.remove(mail['file'])
 
     time.sleep(10)
